Dispatch classes from tools.py in submodule task
  - Manifold is renamed Task
  - COM, EndEffector, Foot, Grasp, OpFrame, Posture, Pregrasp, and Task
    are defined in separate files.
