+++ src/agimus_sot/task/com.py
@@ -1,0 +1,56 @@
+# Copyright 2018, 2019 CNRS - Airbus SAS
+# Author: Joseph Mirabel and Alexis Nicolin
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met:
+
+# 1. Redistributions of source code must retain the above copyright
+# notice, this list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright
+# notice, this list of conditions and the following disclaimer in the
+# documentation and/or other materials provided with the distribution.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+# HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+# SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+# LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+# DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+# THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+from dynamic_graph.sot.core.meta_tasks_kine import MetaTaskKineCom
+from .task import Task
+
+class COM (Task):
+    sep = "_com_"
+    def __init__ (self, comname, sotrobot, withDerivative = False):
+        self.taskCom = MetaTaskKineCom (sotrobot.dynamic,
+                name = COM.sep + comname)
+        super(COM, self).__init__()
+        self.taskCom.task.setWithDerivative (withDerivative)
+        self.tasks = [ self.taskCom.task, ]
+        self.topics[comname] = {
+                "velocity": False,
+                "type": "vector3",
+                "handler": "hppcom",
+                "hppcom": comname,
+                "signalGetters": frozenset([ self._signalPositionRef ])
+                }
+        if withDerivative:
+            self.topics["vel_" + comname] = {
+                    "velocity": True,
+                    "type": "vector3",
+                    "handler": "hppcom",
+                    "hppcom": comname,
+                    "signalGetters": frozenset([ self._signalVelocityRef ])
+                    }
+        self.taskCom.task.controlGain.value = 5
+
+    def _signalPositionRef (self): return self.taskCom.featureDes.errorIN
+    def _signalVelocityRef (self): return self.taskCom.featureDes.errordotIN
